libs/shared-db/pyproject.toml
[project]
name = "shared-db"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "databricks-sdk>=0.25",
  "databricks-sql-connector>=3.3",
  "cachetools>=5.3",
  "slowapi>=0.1.9",
]

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared_db"]
